+++ protocols/dcutr/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.11.0 - unreleased
 
+- Add `ConnectionId` to `Event::DirectConnectionUpgradeSucceeded` and `Event::DirectConnectionUpgradeFailed`.
+  See [PR 4558].
+
+[PR 4558]: https://github.com/libp2p/rust-libp2p/pull/4558
 
 ## 0.10.0